Refactor and enhance tower, GUI and functionality (#61)
* Refactor pixel texture usage to utilize Primitives.Pixel across multiple classes

* Refactor font usage to utilize Primitives.Font across multiple screens and UI components

* Add level label drawing to BaseTower and its derived classes

* Rename DrawLevelLabel to DrawLevelStripes and update its implementation across BaseTower, CannonTower, and MusketTower. Centered DrawLevelStripes position dynamically

* Add inverted button texture and update BuildMenu to use it

* Add cannon and musket icons and update BuildMenu to use new textures

* Enhance CannonProjectile and Explosion handling with AoE damage support and refactor callbacks

* Add preview functionality to UpgradeMenu for tower selection and update drawing logic

* Refactor text color usage across UI components to use Primitives.FontColor for consistency

* dotnet format whitespaces

* remove pixel

* Refactor cannonProjectile onhit, health and damage types to use float for better precision across entities.
aoe damage is based on distance between explosion center and enemy.

* update tower sprites

* musket slight sprite update

* Update cannon tower sprite dimensions and adjust level stripe positioning

* Add FisherTower implementation and integrate into build system

- Introduced FisherTower class inheriting from BaseTower with specific attributes and drawing logic.
- Updated BuildManager to handle FisherTower placement.
- Added FisherTower to BuildMode enum and TowerFactory.
- Integrated FisherTower into GameScreen for placement and selection.
- Updated BuildMenu to include FisherTower button with appropriate cost and selection logic.
- Modified HUD and UpgradeMenu to display FisherTower information correctly.

---------

+++ RumDefence/Gameplay/Systems/LevelProgressSystem.cs
@@ -13,7 +13,7 @@ public class LevelProgressSystem : IGameLoopSystem
     /// <summary>
     /// Indicates the number of lives the player has remaining.
     /// </summary>
-    public int LivesRemaining { get; private set; }
+    public float LivesRemaining { get; private set; }
 
     /// <summary>
     /// Indicates the total number of lives the player had at the start of the level.
@@ -56,7 +56,7 @@ public class LevelProgressSystem : IGameLoopSystem
     /// </summary>
     /// <param name="hits">Amount of hits to take, hits must be a positive integer</param>
     /// <exception cref="ArgumentException">Thrown when the amount of hits is negative</exception>
-    public void TakeHits(int hits)
+    public void TakeHits(float hits)
     {
         if (hits < 0)
         {
